Moved list deal commands into a new deals subcommand
